🔨 refactor(api.ts): refactor api module to use a singleton pattern for ApiService
🔀 chore(api.type.ts): update IApi interface to include patch method for ApiService

The `api.ts` file has been refactored to use a singleton pattern for the `ApiService` class. This ensures that only one instance of the `ApiService` class is created and used throughout the application. The `getInstance` method has been updated to return the instance of the `ApiService` class.

The `api.type.ts` file has been updated to include the `patch` method in the `IApiService` interface. This allows the `ApiService` class to support the `patch` HTTP method.

These changes improve the code structure and maintainability of the API module.